+++ tests/UnityBestPractices.Analyzers.Tests/Program.cs
@@ -279,6 +279,7 @@ internal sealed partial class AnalyzerTests
                 public EntitiesBuilder WithEntityQueryOptions(EntityQueryOptions options) => this;
                 public EntitiesBuilder WithStructuralChanges() => this;
                 public EntitiesBuilder WithoutBurst() => this;
+                public EntitiesBuilder WithReadOnly<T>(T value) where T : struct => this;
                 public ForEachDescription ForEach(EntityAction action) => default;
                 public ForEachDescription ForEach<T>(RefAction<T> action) where T : struct, IComponentData => default;
                 public ForEachDescription ForEach<T>(InAction<T> action) where T : struct, IComponentData => default;